perf: combine api server and relay
