+++ Dockerfile
@@ -103,6 +103,16 @@ RUN --mount=type=cache,target=/go/pkg/mod \
     --mount=type=cache,target=/root/.cache/go-build \
     make build
 
+# Combined API Server and Relay build stage
+FROM common-builder AS apiserver-and-relay-builder
+WORKDIR /app/disperser
+RUN --mount=type=cache,target=/go/pkg/mod \
+    --mount=type=cache,target=/root/.cache/go-build \
+    go build -ldflags="-X main.version=${SEMVER} \
+                       -X main.gitCommit=${GITCOMMIT} \
+                       -X main.gitDate=${GITDATE}" \
+      -o ./bin/combined/apiserver_and_relay ./cmd/apiserver_and_relay
+
 # Final stages for each component
 FROM alpine:3.22 AS churner
 COPY --from=churner-builder /app/operators/bin/churner /usr/local/bin
@@ -151,3 +161,7 @@ ENTRYPOINT ["generator"]
 FROM alpine:3.22 AS generator2
 COPY --from=generator2-builder /app/test/v2/bin/load /usr/local/bin
 ENTRYPOINT ["load", "-", "-"]
+
+FROM alpine:3.18 AS apiserver-and-relay
+COPY --from=apiserver-and-relay-builder /app/disperser/bin/combined/apiserver_and_relay /usr/local/bin
+ENTRYPOINT ["apiserver_and_relay"]
